WIP: Refactoring FieldCollector to get field from schema type
Not sure if we should be doing any of this right now? Field arguments are not necessary in the current implementation because we just use the field name and let you set the value for the field. Since once you convert to a mock you only have one value for the field with a set of arguments you don’t need this robust field argument logic. For fields with different sets of arguments on an operation, you have to provide aliases, which get transformed into their own `MockFields`.